Fix "auto" outline with padding and overflowing inline descendants
This patch fixes a regression in LayoutNG compared to legacy,
when the descendant of an element with padding and "outline: auto"
is an inline block.
And also when it's an inline element (which was not working properly
in legacy layout either, that's why outline-025 and outline-026 tests
fail on legacy).

We need to rebaseline an old pixel test, as we're coming back
to the legacy layout behavior for it.

BUG=1078539
TEST=external/wpt/css/css-ui/outline-024.html
TEST=external/wpt/css/css-ui/outline-025.html
TEST=external/wpt/css/css-ui/outline-026.html
Change-Id: I8ccf72fb34e1d9ce6821eec9312f1fecd52e1281