Improve webxr-test-api polyfill logic
Now that Chromium no longer needs the polyfill to be created before the
xr object is utilized, we can convert the loose promises into promise
functions. This helps by allowing us to remove a duplicate check on the
engine types in the relevant promises, and helps to make the loading
behavior a bit more consistent.

In addition, now that navigator.xr.test can be queried before creating
the polyfill object, simplify the logic in webxr-test.js; this includes
removing a no-longer-used guard for the old WebVR tests.

Fixed:1074452
Change-Id: Ieb52c412a0a4825387ef0b136cc4a85bbdb33191